+++ src/features/auth/index.ts
@@ -1,0 +1,3 @@
+export * from "./context/AuthProvider";
+export { default as LoginPage } from "./pages/LoginPage";
+export { default as RegisterPage } from "./pages/RegisterPage";
