+++ src/features/auth/routes.tsx
@@ -1,6 +1,5 @@
 import { Route } from "react-router-dom";
-import LoginPage from "./pages/LoginPage";
-import RegisterPage from "./pages/RegisterPage";
+import { LoginPage, RegisterPage } from "./index";
 
 export const AuthRoutes = (
   <>
