+++ src/features/games/index.ts
@@ -1,0 +1,7 @@
+export * from "./hooks/gameQueries";
+export * from "./hooks/useSearchQueries";
+export * from "./hooks/useGameListQueries";
+export { default as GameDetailPage } from "./pages/GameDetailPage";
+export { default as GameListPage } from "./pages/GameListPage";
+export { default as HomePage } from "./pages/HomePage";
+export { default as SearchEnginePage } from "./pages/SearchEnginePage";
