+++ src/features/misc/index.ts
@@ -1,0 +1,2 @@
+export { default as StartPage } from "./pages/StartPage";
+export { default as NotFound } from "./pages/NotFound";
