+++ src/features/notifications/index.ts
@@ -1,0 +1,2 @@
+export * from "./hooks/notificationQueries";
+export { default as NotificationsPage } from "./pages/NotificationsPage";
