+++ src/features/notifications/routes.tsx
@@ -1,5 +1,5 @@
 import { Route } from "react-router-dom";
-import NotificationsPage from "./pages/NotificationsPage";
+import { NotificationsPage } from "./index";
 import RequireAuth from "@/features/auth/components/RequireAuth";
 
 export const NotificationRoutes = (
