+++ src/features/users/index.ts
@@ -1,0 +1,4 @@
+export * from "./hooks/userQueries";
+export * from "./hooks/friendshipQueries";
+export { default as UserProfilePage } from "./pages/UserProfilePage";
+export { default as UserSettingsPage } from "./pages/UserSettingsPage";
